# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Marketplace API",
    "description": "API for Village Website & Marketplace"
  },
  "components": {
    "schemas": {
...
      "CreateProducePayload": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Organic Honeycrisp Apples",
            "description": "The public title of the produce listing"
          },
          "produceType": {
            "type": "string",
            "example": "fruit",
            "description": "Category of the produce (e.g., vegetable, fruit, herb)"
          },
          "pricePerOz": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true,
            "example": 12.5,
            "description": "Price/Amount in USD"
          },
          "totalOzInventory": {
            "type": "number",
            "minimum": 0,
            "example": 16,
            "description": "Weight in ounces (oz)"
          },
          "availableBy": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": "2026-03-25T10:00:00Z",
            "description": "The date and time when the produce will be ready for pickup or delivery"
          },
          "harvestFrequencyDays": {
            "type": "integer",
            "minimum": 0,
            "example": 7,
            "description": "How often this item is typically harvested, in days"
          },
          "seasonStart": {
            "type": "string",
            "format": "date",
            "example": "2026-03-27",
            "description": "ISO 8601 calendar date (YYYY-MM-DD)"
          },
          "seasonEnd": {
            "type": "string",
            "format": "date",
            "example": "2026-03-27",
            "description": "ISO 8601 calendar date (YYYY-MM-DD)"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri",
              "example": "https://assets.example.com/produce/apples.jpg",
              "description": "A fully qualified URL to an image asset"
            },
            "default": [],
            "example": ["https://example.com/apple.jpg"],
            "description": "List of gallery images for the product"
          },
          "isSubscribable": {
            "type": "boolean",
            "default": false,
            "example": true,
            "description": "Whether customers can set up recurring orders for this item"
          }
        },
        "required": [
          "title",
          "pricePerOz",
          "totalOzInventory",
          "harvestFrequencyDays",
          "seasonStart",
          "seasonEnd"
        ]
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "buyerId": {
            "type": "string"
          },
          "sellerId": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string",
            "enum": ["card", "snap"]
          },
          "fulfillmentType": {
            "type": "string",
            "enum": ["pickup", "delivery"]
          },
          "scheduledTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true,
            "enum": ["pending", "completed", "canceled", null]
          },
          "cancelReason": {
            "type": "string",
            "nullable": true
          },
          "totalAmount": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "buyerId",
          "sellerId",
          "paymentMethod",
          "fulfillmentType",
          "scheduledTime",
          "status",
          "cancelReason",
          "totalAmount",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateReview": {
        "type": "object",
        "properties": {
          "sellerId": {
            "type": "string",
            "example": "user_123456789",
            "description": "The unique identifier for the user, prefixed with \"user_\""
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "The standard UUID identifier for the resource"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5,
            "description": "Star rating from 1 to 5"
          },
          "comment": {
            "type": "string",
            "example": "The apples were incredibly crisp and sweet!",
            "description": "Optional text review from the buyer"
          }
        },
        "required": ["sellerId", "orderId", "rating"]
      }
    },
...
  },
  "paths": {
...
    "/api/produce": {
      "post": {
        "operationId": "createProduce",
        "description": "Create a new produce listing.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProducePayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart": {
      "get": {
        "operationId": "getCart",
        "description": "Fetch user's active cart, grouped by seller. Drops expired reservations automatically.",
        "responses": {
          "200": {
            "description": "Cart Object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cart": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "seller": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "example": "user_123456789",
                                "description": "The unique identifier for the user, prefixed with \"user_\""
                              },
                              "name": {
                                "type": "string",
                                "nullable": true,
                                "example": "Sun-Kissed Orchards",
                                "description": "The name of the seller providing these items"
                              }
                            },
                            "required": ["id", "name"]
                          },
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "reservationId": {
                                  "type": "string",
                                  "format": "uuid",
                                  "example": "123e4567-e89b-12d3-a456-426614174000",
                                  "description": "The standard UUID identifier for the resource"
                                },
                                "productId": {
                                  "type": "string",
                                  "format": "uuid",
                                  "example": "123e4567-e89b-12d3-a456-426614174000",
                                  "description": "The standard UUID identifier for the resource"
                                },
                                "title": {
                                  "type": "string",
                                  "example": "Heirloom Tomatoes",
                                  "description": "Display title of the product"
                                },
                                "pricePerOz": {
                                  "type": "string",
                                  "example": "0.50",
                                  "description": "Price per ounce formatted as a string"
                                },
                                "quantityOz": {
                                  "type": "string",
                                  "example": "16.0",
                                  "description": "Quantity in ounces currently in cart"
                                },
                                "isSubscription": {
                                  "type": "boolean",
                                  "nullable": true,
                                  "example": false,
                                  "description": "Flag indicating if this is a subscription item"
                                },
                                "expiresAt": {
                                  "type": "string",
                                  "format": "date-time",
                                  "example": "2026-03-27T09:00:00Z",
                                  "description": "ISO 8601 UTC timestamp"
                                },
                                "images": {
                                  "type": "array",
                                  "nullable": true,
                                  "items": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://assets.example.com/produce/apples.jpg",
                                    "description": "A fully qualified URL to an image asset"
                                  }
                                }
                              },
                              "required": [
                                "reservationId",
                                "productId",
                                "title",
                                "pricePerOz",
                                "quantityOz",
                                "isSubscription",
                                "expiresAt",
                                "images"
                              ]
                            },
                            "description": "List of items in the cart belonging to this specific seller"
                          }
                        },
                        "required": ["seller", "items"]
                      },
                      "description": "User shopping cart grouped by seller"
                    }
                  },
                  "required": ["cart"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "operationId": "getOrders",
        "description": "Get historical or active orders.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["buyer", "seller"],
              "description": "The perspective from which to fetch orders",
              "example": "buyer"
            },
            "required": true,
            "description": "The perspective from which to fetch orders",
            "name": "role",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["pending", "completed", "canceled"],
              "example": "pending",
              "description": "The lifecycle status of the order"
            },
            "required": false,
            "description": "The lifecycle status of the order",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Optional timeframe filter (e.g., \"7d\", \"30d\", or ISO range)",
              "example": "recent"
            },
            "required": false,
            "description": "Optional timeframe filter (e.g., \"7d\", \"30d\", or ISO range)",
            "name": "timeframe",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/messages": {
      "get": {
        "operationId": "getMessages",
        "description": "Short-polling endpoint for the active chat screen.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "conversationId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": true,
            "name": "since",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Messages list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "nullable": true
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "sendMessage",
        "description": "Send a message.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                },
                "required": ["conversationId", "text"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": ["success"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reviews": {
      "post": {
        "operationId": "createReview",
        "description": "Leave a star rating and optional comment for a completed order.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReview"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "reviewId": {
                      "type": "string"
                    }
                  },
                  "required": ["success", "reviewId"]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., review already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contact": {
      "post": {
        "operationId": "submitContactForm",
        "description": "Submit a general contact form. Forwards the message and sends an auto-reply.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Jane Doe",
                    "description": "Full name of the person submitting the inquiry"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "jane@example.com",
                    "description": "Contact email address"
                  },
                  "company": {
                    "type": "string",
                    "example": "Fresh Eats LLC",
                    "description": "Optional company or organization name"
                  },
                  "message": {
                    "type": "string",
                    "minLength": 1,
                    "example": "I am interested in bulk purchasing.",
                    "description": "The content of the contact request"
                  }
                },
                "required": ["name", "email", "message"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact form submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": ["success"]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    }
  }
}